chore(frontend): routing refactor, debt allowance
