+++ packages/frontend/components/Borrow/TransactionModal.tsx
@@ -36,6 +36,7 @@ import AddTokenButton from "../Shared/AddTokenButton"
 import { showPosition } from "../../helpers/navigation"
 import { useRouter } from "next/router"
 import { vaultFromAddress } from "../../helpers/positions"
+import { camelize } from "../../helpers/values"
 
 type InvalidStep = {
   label: "Invalid"
@@ -101,18 +102,20 @@ function TransactionModal({ hash }: TransactionModalProps) {
         zIndex: 1,
       }
 
-      const label =
+      const action = step.toString()
+      const preposition =
         step === RoutingStep.DEPOSIT
-          ? `Deposit ${amount} ${token.symbol} on ${provider}`
-          : step === RoutingStep.BORROW
-          ? `Borrow ${amount} ${token.symbol} from ${provider}`
-          : step === RoutingStep.WITHDRAW
-          ? `Withdraw ${amount} ${token.symbol} from ${provider}`
-          : step === RoutingStep.PAYBACK
-          ? `Repay ${amount} ${token.symbol} from ${provider}`
+          ? "on"
           : step === RoutingStep.X_TRANSFER
-          ? `Bridge ${amount} ${token.symbol} to ${chain}`
-          : "Invalid"
+          ? "to"
+          : "from"
+
+      const label =
+        step === RoutingStep.START || step === RoutingStep.END
+          ? "Invalid"
+          : camelize(
+              `${action} ${amount} ${token.symbol} ${preposition} ${provider}`
+            )
 
       const icon =
         step === RoutingStep.X_TRANSFER ? (
