+++ packages/frontend/helpers/history.ts
@@ -1,56 +1,23 @@
-import { ChainId, RoutingStep } from "@x-fuji/sdk"
+import { ChainId, RoutingStepDetails } from "@x-fuji/sdk"
 import { formatUnits } from "ethers/lib/utils"
-import { HistoryRoutingStep } from "../store/history.store"
 import { camelize } from "./values"
 
-const stepOutput = (
-  step: HistoryRoutingStep | undefined
-): string | undefined => {
-  if (!step) return undefined
-  const stepAction =
-    step.step === RoutingStep.DEPOSIT
-      ? "deposit"
-      : step.step === RoutingStep.BORROW
-      ? "borrow"
-      : step.step === RoutingStep.PAYBACK
-      ? "payback"
-      : "withdraw"
-  const stepAmount =
-    step?.amount && formatUnits(step.amount, step.token.decimals)
-
-  return `${stepAction} ${stepAmount} ${step.token.symbol}}`
-}
-
 export const entryOutput = (
-  steps: HistoryRoutingStep[]
+  step: RoutingStepDetails,
+  hash: string
 ): {
   title: string
   transactionUrl: {
     hash: string
-    chainId: ChainId | undefined
+    chainId: ChainId
   }
 } => {
-  const step1 =
-    steps.find((s) => s.step === RoutingStep.DEPOSIT) ||
-    steps.find((s) => s.step === RoutingStep.PAYBACK)
+  const stepAction = camelize(step.step.toString())
+  const stepAmount =
+    step.amount && formatUnits(step.amount, step.token?.decimals)
+  const title = `${stepAction} ${stepAmount} ${step.token?.symbol}}`
+  const chainId = step.chainId
 
-  const step1Output = stepOutput(step1)
-
-  const step2 =
-    steps.find((s) => s.step === RoutingStep.BORROW) ||
-    steps.find((s) => s.step === RoutingStep.WITHDRAW)
-  const step2Output = stepOutput(step2)
-
-  const title = camelize(
-    step1Output
-      ? step2Output
-        ? `${step1Output} and ${step2Output}`
-        : `${step1Output}`
-      : `${step2Output}`
-  )
-
-  const hash = step2?.txHash || step2?.txHash || ""
-  const chainId = step2?.chainId || step2?.chainId
   return {
     title,
     transactionUrl: {
